Update authentication.md
I ran into this error "Cannot read properties of null (reading 'id')" a few times when using currentUser on the web side, and it also seems to be an error a few other people have faced/posted about, something in the tutorial addressing it seems like it would be good. This is the second PR solving the same issue, as I submitted one adding a callout/note but was recommended by Pan to directly add it to the code in the tutorial as people might copy and paste and miss the call-out, which I also think is a good reason to directly add.